requirements.txt
alembic==1.10.2
asyncpg==0.27.0
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.6
python-dotenv==1.0.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
websockets==10.4
